requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
tqdm==4.66.1
